Created Ability Class
Created the Ability class, which I didn't even have in my UML. Will add
it. It has a name, description, type, which is either physical or
magical, which indicates which stats to use to calculate its final
damage, target spread, which indicates if it hits the entire party or a
single target,  affect enemy, which indicates if it's an ability to
affect the enemy or the party, and base stat, which indicates how
powerful the ability is and how much damage or healing it will do.

Added the getCharacter method to Party, as we indicated we would in our
master UML.