Complete Dungeon and Battle Loop
Complete how the game loop is going to function. The Dungeon is key. It
gets input from the view, updates its data, then sees if a battle has
occurred. If it has, it calls the battle to do its work, and waits, then
it goes back to the run loop in game, and sees if the party lived or
killed the dragon. If the party has lived and not killed the dragon, the
process repeats, if the party died or if the dragon died, then the game
is over and the loop breaks.