CSGameSystem: Code Consolidation attempt
I have combined all the classes from all the projects in the GitHub
repository,
to form a single Visual Studio Solution called CSGameSystem.

I got all of the code to compile, but some of it needed to be commented
out.

I think we need to be all working from a single Visual Studio solution,
so that
we have access to all the code.