Created Character UML and made changes across the System.
Enemies and their AI now return a BattleEvent, which we called Event in
UML, but that's a reserved word in C#, so that the Battle System knows
what choices it made. I also changed the BaseStat property of Ability to
BaseDamage because it made more sense to me. Added a Null Object to
weapons, which should never happen, but just incase, it's there. Altered
the way level is determined. For now, every 1000 experience is a level.
Finally, created new Character System UML in Umlet, which is much nicer
looking, though it contains none of my future plans.