Character Factory Finish
Got the Character Factory for initial PlayerCharacter creation finished.
Changed the ArmorFactory and the CharacterFactory to singleton. Got rid
of the NULL value in all enumerations, as it was unnecessary. Added an
isDead property to the Party object that'll tell you if anyone in the
party is still alive.  Update New Character UML as necessary. Added in
the LevelTracts for each class, though they're not complete in terms of
implementation.