Finished the Enemy Factory. Items are all that are left.
Moved the EnemyType enum from SystemEnum to CharacterEnumerations. The
Dragon value is just there to indicate to the view that it is the
dragon. Other than that, it serves no purpose, as there is a separate
function for getting the Dragon. It is used to get specific enemies
within the factory, which is based off the party's highest level
character.