Fixed A lot of Things
Characters compareTo was backwards. Fixed that.
Added notifyGameOver in text view as well as when something dies. Added
code for updating victory and notifying when the party is dead through
out game, view, and battle.
Added accessor properties in BattleEvent.
Altered the stats in EnemyFactory just for the sake of testing.
Fixed constants in Ability. Healing abilities we damaging and vice verse
sa.
Use the name property for abilities in AbilityAction, since abilites
don't have a to string.
Fixed Healer Attacker AI to attack only alive characters. This is a
problem through out all of my AI classes. Completely forgot to check for
that.