Updated Numerous Things
The Battle now notifies when the party has died or when the dragon has
died for the dragon battle.
Removed all stat alteration in each class' level tract since that is
also instantiated in the character factory.
Fixed my misspelling of thief.
Updated Dungeon with Robert's hard coded grid.
Added notifiyUlitimate victory to the View interface so that the View
can say that the party won.
Problems:
The Dragon is murdering itself. I don't know why. Could be an AI problem
or an AbilityAction problem.